pyproject.toml
[tool.poetry]
name = "facial-r"
version = "0.1.0"
description = "A face recognition application using FastAPI and Firebase."
authors = ["chukwumdiebube ojinta <chukwumdiebubeojinta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
numpy = "^1.21.2"
cmake = "^3.30.0"
dlib = "^19.24.4"
face-recognition = "^1.3.0"
cvzone = "^1.6.1"
opencv-python = "4.5.4.60"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
fastapi-socketio = "^0.0.10"
pillow = "^10.4.0"
firebase-admin = "^6.5.0"
pydantic-settings = "^2.3.4"
pytest = "^8.3.1"
python-socketio = {extras = ["asgi"], version = "^5.11.3"}
pytest-mock = "^3.14.0"
google-cloud-storage = "^2.18.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
